+++ changelog.md
@@ -1,3 +1,21 @@
+# <sub>v2.0.0-alpha</sub>
+#### _Nov. 26, 2017_
+
+ * Rename `isValidDateString` to `isDateString`.
+ * Rename `isValidId` to `isId`.
+ * Rename `isAllValidId` to `everyIsId`.
+ * Rename `isAllValidUniqueId` to `everyIsUniqueId`.
+ * Rename `isValidObjectId` to `isObjectId`.
+ * Rename `isAllValidObjectId` to `everyIsObjectId`.
+ * Rename `isAllValidUniqueObjectId` to `everyIsUniqueObjectId`.
+ * Rename `isAllValidObjectIdOrNull` to `everyIsObjectIdOrNull`.
+ * Rename `isAllValidUniqueObjectIdOrNull` to `everyIsUniqueObjectIdOrNull`.
+ * Rename `isValidEmail` to `isEmail`.
+ * Rename `isValidPhoneSimple` to `isSimplePhoneNumber`.
+ * Rename `isAllAllowed` to `everyIsAllowed`.
+ * Rename `isAllUniqueAllowed` to `everyIsUniqueAllowed`.
+ * Rename `isAllowedStringFields` to `isFieldsString`.
+ 
 # <sub>v1.1.0</sub>
 #### _Nov. 26, 2017_